Draft: Accommodate calendar-specific fields in non-ISO ...FromFields() methods
Previously, the spec hardcoded "era" and "eraYear" as extra fields that
should be accessed on the property bags passed to the dateFromFields(),
yearMonthFromFields(), and monthDayFromFields() methods of
Temporal.Calendar.

Instead, the extra fields should be obtained by calling the calendar's
fields() method. This accommodates calendars that require additional extra
fields, beyond "era" and "eraYear".

I'm not sure whether this is ultimately a normative change or not, but if
it, is then we can also fix an observable inconsistency in the order of
operations between the three ...FromFields() methods, while we are here
anyway. (If not, then revert that line before submitting this.)

Closes: #1789